Don't attempt to loop over null lists in {AP,VA}Model.notifyDeleteDraftVersion()
Sigh, fix two more occurrences of the error in which I loop over a
list without first checking if the list is null.

Fix the
{AccessPoints,VersionArtefacts}Model.notifyDeleteDraftVersion()
methods.

+++ src/main/java/au/org/ands/vocabs/registry/model/AccessPointsModel.java
@@ -407,13 +407,16 @@ public class AccessPointsModel extends ModelBase {
     /** {@inheritDoc} */
     @Override
     protected void notifyDeleteDraftVersion(final Integer versionId) {
-        for (AccessPoint ap : draftAPs.get(versionId)) {
-            // Just delete the row;
-            // for a draft instance, no workflow is applied.
-            AccessPointDAO.deleteAccessPoint(em(), ap);
+        List<AccessPoint> draftAPList = draftAPs.get(versionId);
+        if (draftAPList != null) {
+            for (AccessPoint ap : draftAPs.get(versionId)) {
+                // Just delete the row;
+                // for a draft instance, no workflow is applied.
+                AccessPointDAO.deleteAccessPoint(em(), ap);
+            }
+            // Remove from our own records.
+            draftAPs.remove(versionId);
         }
-        // Remove from our own records.
-        draftAPs.remove(versionId);
     }
 
     /** {@inheritDoc}
